refactor(plc): extract PlcQuizLibraryBody + PlcAssignmentsBody (Phase 2B)
Mirror the NotesBody / TodosBody / MembersBody pattern from #1582: the
two remaining tab bodies move into components/plc/bodies/ so the v2 grid
renderer (and the fullscreen expansion path in
PlcDashboard.renderExpandedBody) can mount them without going through
tab chrome. The tab files become 3-line wrappers used only by the
legacy v1 dashboard tab routing.

No behavior change.
